Code review fixes #1
