+++ distribution/services/create.go
@@ -26,28 +26,29 @@ func (cb *CreateReleaseBundleService) GetDistDetails() auth.CommonDetails {
 	return cb.DistDetails
 }
 
-func (cb *CreateReleaseBundleService) CreateReleaseBundle(createBundleParams CreateUpdateReleaseBundleParams) error {
-	CreateUpdateReleaseBundleBody, err := CreateBundleBody(createBundleParams, cb.DryRun)
+func (cb *CreateReleaseBundleService) CreateReleaseBundle(createBundleParams CreateReleaseBundleParams) error {
+	releaseBundleBody, err := distrbutionServiceUtils.CreateBundleBody(createBundleParams.ReleaseBundleParams, cb.DryRun)
 	if err != nil {
 		return err
 	}
-	createReleaseBundleBody := &CreateReleaseBundleBody{
-		Name:              createBundleParams.Name,
-		Version:           createBundleParams.Version,
-		ReleaseBundleBody: *CreateUpdateReleaseBundleBody,
+	
+	body := &createReleaseBundleBody{
+		Name: createBundleParams.Name,
+		Version: createBundleParams.Version,
+		ReleaseBundleBody: *releaseBundleBody,
 	}
 
-	return cb.execCreateReleaseBundle(createBundleParams.GpgPassphrase, createReleaseBundleBody)
+	return cb.execCreateReleaseBundle(createBundleParams.GpgPassphrase, body)
 }
 
-func (cb *CreateReleaseBundleService) execCreateReleaseBundle(gpgPassprase string, releaseBundle *CreateReleaseBundleBody) error {
+func (cb *CreateReleaseBundleService) execCreateReleaseBundle(gpgPassprase string, releaseBundle *createReleaseBundleBody) error {
 	httpClientsDetails := cb.DistDetails.CreateHttpClientDetails()
 	content, err := json.Marshal(releaseBundle)
 	if err != nil {
 		return errorutils.CheckError(err)
 	}
 	url := cb.DistDetails.GetUrl() + "api/v1/release_bundle"
-	distrbutionServiceUtils.SetGpgPassphrase(gpgPassprase, &httpClientsDetails.Headers)
+	distrbutionServiceUtils.AddGpgPassphraseHeader(gpgPassprase, &httpClientsDetails.Headers)
 	artifactoryUtils.SetContentType("application/json", &httpClientsDetails.Headers)
 	resp, body, err := cb.client.SendPost(url, content, &httpClientsDetails)
 	if err != nil {
@@ -57,13 +58,26 @@ func (cb *CreateReleaseBundleService) execCreateReleaseBundle(gpgPassprase strin
 		return errorutils.CheckError(errors.New("Distribution response: " + resp.Status + "\n" + utils.IndentJson(body)))
 	}
 
-	log.Debug("Artifactory response: ", resp.Status)
-	log.Output(utils.IndentJson(body))
+	log.Debug("Distribution response: ", resp.Status)
+	log.Debug(utils.IndentJson(body))
 	return errorutils.CheckError(err)
 }
 
-type CreateReleaseBundleBody struct {
+type createReleaseBundleBody struct {
 	Name    string `json:"name"`
 	Version string `json:"version"`
-	ReleaseBundleBody
+	distrbutionServiceUtils.ReleaseBundleBody
+}
+
+type CreateReleaseBundleParams struct {
+	distrbutionServiceUtils.ReleaseBundleParams
+}
+
+func NewCreateReleaseBundleParams(name, version string) CreateReleaseBundleParams {
+	return CreateReleaseBundleParams{
+		distrbutionServiceUtils.ReleaseBundleParams{
+			Name:    name,
+			Version: version,
+		},
+	}
 }
