+++ distribution/services/distribute.go
@@ -63,7 +63,7 @@ func (dr *DistributeReleaseBundleService) execDistribute(name, version string, d
 	}
 
 	log.Debug("Distribution response: ", resp.Status)
-	log.Output(utils.IndentJson(body))
+	log.Debug(utils.IndentJson(body))
 	return errorutils.CheckError(err)
 }
