+++ distribution/services/update.go
@@ -14,14 +14,6 @@ import (
 	"github.com/jfrog/jfrog-client-go/utils/log"
 )
 
-type ReleaseNotesSyntax string
-
-const (
-	Markdown  ReleaseNotesSyntax = "markdown"
-	Asciidoc                     = "asciidoc"
-	PlainText                    = "plain_text"
-)
-
 type UpdateReleaseBundleService struct {
 	client      *rthttpclient.ArtifactoryHttpClient
 	DistDetails auth.CommonDetails
@@ -36,22 +28,22 @@ func (ur *UpdateReleaseBundleService) GetDistDetails() auth.CommonDetails {
 	return ur.DistDetails
 }
 
-func (ur *UpdateReleaseBundleService) UpdateReleaseBundle(createBundleParams CreateUpdateReleaseBundleParams) error {
-	releaseBundleBody, err := CreateBundleBody(createBundleParams, ur.DryRun)
+func (ur *UpdateReleaseBundleService) UpdateReleaseBundle(createBundleParams UpdateReleaseBundleParams) error {
+	releaseBundleBody, err := distrbutionServiceUtils.CreateBundleBody(createBundleParams.ReleaseBundleParams, ur.DryRun)
 	if err != nil {
 		return err
 	}
 	return ur.execUpdateReleaseBundle(createBundleParams.Name, createBundleParams.Version, createBundleParams.GpgPassphrase, releaseBundleBody)
 }
 
-func (ur *UpdateReleaseBundleService) execUpdateReleaseBundle(name, version, gpgPassphrase string, releaseBundle *ReleaseBundleBody) error {
+func (ur *UpdateReleaseBundleService) execUpdateReleaseBundle(name, version, gpgPassphrase string, releaseBundle *distrbutionServiceUtils.ReleaseBundleBody) error {
 	httpClientsDetails := ur.DistDetails.CreateHttpClientDetails()
 	content, err := json.Marshal(releaseBundle)
 	if err != nil {
 		return errorutils.CheckError(err)
 	}
 	url := ur.DistDetails.GetUrl() + "api/v1/release_bundle/" + name + "/" + version
-	distrbutionServiceUtils.SetGpgPassphrase(gpgPassphrase, &httpClientsDetails.Headers)
+	distrbutionServiceUtils.AddGpgPassphraseHeader(gpgPassphrase, &httpClientsDetails.Headers)
 	artifactoryUtils.SetContentType("application/json", &httpClientsDetails.Headers)
 	resp, body, err := ur.client.SendPut(url, content, &httpClientsDetails)
 	if err != nil {
@@ -61,85 +53,19 @@ func (ur *UpdateReleaseBundleService) execUpdateReleaseBundle(name, version, gpg
 		return errorutils.CheckError(errors.New("Distribution response: " + resp.Status + "\n" + utils.IndentJson(body)))
 	}
 
-	log.Debug("Artifactory response: ", resp.Status)
-	log.Output(utils.IndentJson(body))
+	log.Debug("Distribution response: ", resp.Status)
 	return errorutils.CheckError(err)
 }
 
-func CreateBundleBody(createBundleParams CreateUpdateReleaseBundleParams, dryRun bool) (*ReleaseBundleBody, error) {
-	var bundleQueries []BundleQuery
-	// Create release bundle queries
-	for _, specFile := range createBundleParams.SpecFiles {
-		if specFile.GetSpecType() != artifactoryUtils.AQL {
-			query, err := artifactoryUtils.CreateAqlBodyForSpecWithPattern(specFile)
-			if err != nil {
-				return nil, err
-			}
-			specFile.Aql = artifactoryUtils.Aql{ItemsFind: query}
-			aql := artifactoryUtils.BuildQueryFromSpecFile(specFile, artifactoryUtils.NONE)
-			bundleQueries = append(bundleQueries, BundleQuery{Aql: aql})
-		}
-	}
+type UpdateReleaseBundleParams struct {
+	distrbutionServiceUtils.ReleaseBundleParams
+}
 
-	// Create release bundle struct
-	releaseBundleBody := &ReleaseBundleBody{
-		DryRun:            dryRun,
-		SignImmediately:   createBundleParams.SignImmediately,
-		StoringRepository: createBundleParams.StoringRepository,
-		Description:       createBundleParams.Description,
-		BundleSpec: BundleSpec{
-			Queries: bundleQueries,
+func NewUpdateReleaseBundleParams(name, version string) UpdateReleaseBundleParams {
+	return UpdateReleaseBundleParams{
+		distrbutionServiceUtils.ReleaseBundleParams{
+			Name:    name,
+			Version: version,
 		},
-	}
-
-	// Add relese notes if needed
-	if createBundleParams.ReleaseNotes != "" {
-		releaseBundleBody.ReleaseNotes = &ReleaseNotes{
-			Syntax:  createBundleParams.ReleaseNotesSyntax,
-			Content: createBundleParams.ReleaseNotes,
-		}
-	}
-	return releaseBundleBody, nil
-}
-
-type ReleaseBundleBody struct {
-	DryRun            bool          `json:"dry_run"`
-	SignImmediately   bool          `json:"sign_immediately,omitempty"`
-	StoringRepository string        `json:"storing_repository,omitempty"`
-	Description       string        `json:"description,omitempty"`
-	ReleaseNotes      *ReleaseNotes `json:"release_notes,omitempty"`
-	BundleSpec        BundleSpec    `json:"spec"`
-}
-
-type ReleaseNotes struct {
-	Syntax  ReleaseNotesSyntax `json:"syntax,omitempty"`
-	Content string             `json:"content,omitempty"`
-}
-
-type BundleSpec struct {
-	Queries []BundleQuery `json:"queries"`
-}
-
-type BundleQuery struct {
-	QueryName string `json:"query_name,omitempty"`
-	Aql       string `json:"aql"`
-}
-
-type CreateUpdateReleaseBundleParams struct {
-	SpecFiles          []*artifactoryUtils.ArtifactoryCommonParams
-	Name               string
-	Version            string
-	SignImmediately    bool
-	StoringRepository  string
-	Description        string
-	ReleaseNotes       string
-	ReleaseNotesSyntax ReleaseNotesSyntax
-	GpgPassphrase      string
-}
-
-func NewCreateUpdateBundleParams(name, version string) CreateUpdateReleaseBundleParams {
-	return CreateUpdateReleaseBundleParams{
-		Name:    name,
-		Version: version,
 	}
 }
